requirements.txt
asn1crypto==1.5.1
boto3==1.38.20
botocore==1.38.20
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
dotenv==0.9.9
filelock==3.18.0
idna==3.10
jmespath==1.0.1
packaging==25.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
pyarrow==20.0.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.12.0
six==1.17.0
snowflake-connector-python==3.15.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
tomlkit==0.13.3
typing_extensions==4.13.2
urllib3==2.4.0
